+++ src/doccmd/__init__.py
@@ -274,15 +274,28 @@ def _map_languages_to_suffix() -> dict[str, str]:
 
 
 @beartype
-def _get_skip_directives(markers: Iterable[str]) -> Sequence[str]:
+def _get_group_directives(markers: Iterable[str]) -> Sequence[str]:
     """
-    Skip directives based on the provided skip markers.
+    Group directives based on the provided markers.
     """
     directives: Sequence[str] = []
 
     for marker in markers:
-        skip_directive = rf"skip doccmd[{marker}]"
-        directives = [*directives, skip_directive]
+        directive = rf"group doccmd[{marker}]"
+        directives = [*directives, directive]
+    return directives
+
+
+@beartype
+def _get_skip_directives(markers: Iterable[str]) -> Sequence[str]:
+    """
+    Skip directives based on the provided markers.
+    """
+    directives: Sequence[str] = []
+
+    for marker in markers:
+        directive = rf"skip doccmd[{marker}]"
+        directives = [*directives, directive]
     return directives
 
 
@@ -399,6 +412,7 @@ def _run_args_against_docs(
     pad_temporary_file: bool,
     verbose: bool,
     skip_markers: Iterable[str],
+    group_markers: Iterable[str],
     use_pty: bool,
     markup_language: MarkupLanguage,
 ) -> None:
@@ -458,7 +472,20 @@ def _run_args_against_docs(
         )
     ]
 
-    parsers: Sequence[Parser] = [*code_block_parsers, *skip_parsers]
+    group_markers = {*group_markers, "all"}
+    group_directives = _get_group_directives(markers=group_markers)
+    group_parsers = [
+        markup_language.group_parser_cls(
+            directive=group_directive,
+            evaluator=evaluator,
+        )
+        for group_directive in group_directives
+    ]
+    parsers: Sequence[Parser] = [
+        *code_block_parsers,
+        *skip_parsers,
+        *group_parsers,
+    ]
     sybil = Sybil(parsers=parsers, encoding=encoding)
 
     document = _parse_file(sybil=sybil, path=document_path)
@@ -538,6 +565,33 @@ def _run_args_against_docs(
         To skip a code block for each of multiple markers, for example to skip
         a code block for the ``type-check`` and ``lint`` markers but not all
         markers, add multiple ``skip doccmd`` comments above the code block.
+        """
+    ),
+    multiple=True,
+    callback=_deduplicate,
+)
+@click.option(
+    "group_markers",
+    "--group-marker",
+    type=str,
+    default=None,
+    show_default=True,
+    required=False,
+    help=(
+        """\
+        The marker used to identify code blocks to be grouped.
+
+        By default, code blocks which come just between comments matching
+        'group doccmd[all]: start' and 'group doccmd[all]: end' are grouped
+        (e.g. `.. group doccmd[all]: start` in reStructuredText, `<!--- group
+        doccmd[all]: start -->` in Markdown, or `% group doccmd[all]: start` in
+        MyST).
+
+        When using this option, those, and code blocks which are grouped by
+        a comment including the given marker are ignored. For example, if the
+        given marker is 'type-check', code blocks which come within comments
+        matching 'group doccmd[type-check]: start' and
+        'group doccmd[type-check]: end' are also skipped.
         """
     ),
     multiple=True,
@@ -697,6 +751,7 @@ def main(
     pad_file: bool,
     verbose: bool,
     skip_markers: Sequence[str],
+    group_markers: Sequence[str],
     use_pty_option: _UsePty,
     rst_suffixes: Sequence[str],
     myst_suffixes: Sequence[str],
@@ -760,6 +815,7 @@ def main(
                     temporary_file_extension=temporary_file_extension,
                     temporary_file_name_prefix=temporary_file_name_prefix,
                     skip_markers=skip_markers,
+                    group_markers=group_markers,
                     use_pty=use_pty,
                     markup_language=markup_language,
                 )